organize folder structure for fhir module
