+++ autotest-plr/src/main/java/ca/bc/gov/health/qa/autotest/plr/fhir/FHIRController.java
@@ -4,19 +4,19 @@ import org.apache.logging.log4j.Logger;
 import org.json.JSONObject;
 
 import ca.bc.gov.health.qa.autotest.plr.fhir.actions.FHIRSession;
-import ca.bc.gov.health.qa.autotest.plr.fhir.data.FacilityDataGenerator;
-import ca.bc.gov.health.qa.autotest.plr.fhir.data.FacilityBuilderFactory;
-import ca.bc.gov.health.qa.autotest.plr.fhir.data.FacilityMaintainConfig;
-import ca.bc.gov.health.qa.autotest.plr.fhir.data.OrganizationBuilderFactory;
-import ca.bc.gov.health.qa.autotest.plr.fhir.data.OrganizationDataGenerator;
-import ca.bc.gov.health.qa.autotest.plr.fhir.data.OrganizationMaintainConfig;
+import ca.bc.gov.health.qa.autotest.plr.fhir.data.facility.FacilityBuilderFactory;
+import ca.bc.gov.health.qa.autotest.plr.fhir.data.facility.FacilityDataGenerator;
+import ca.bc.gov.health.qa.autotest.plr.fhir.data.facility.FacilityMaintainConfig;
+import ca.bc.gov.health.qa.autotest.plr.fhir.data.organization.OrganizationBuilderFactory;
+import ca.bc.gov.health.qa.autotest.plr.fhir.data.organization.OrganizationDataGenerator;
+import ca.bc.gov.health.qa.autotest.plr.fhir.data.organization.OrganizationMaintainConfig;
+import ca.bc.gov.health.qa.autotest.plr.fhir.maintain.facility.FacilityQueryResponseMapper;
+import ca.bc.gov.health.qa.autotest.plr.fhir.maintain.facility.MaintainFacilityBuilder;
+import ca.bc.gov.health.qa.autotest.plr.fhir.maintain.organization.MaintainOrgBuilder;
+import ca.bc.gov.health.qa.autotest.plr.fhir.maintain.organization.OrgQueryResponseMapper;
 import ca.bc.gov.health.qa.autotest.plr.fhir.model.IdentifierType;
 import ca.bc.gov.health.qa.autotest.plr.fhir.model.OrgRoleType;
 import ca.bc.gov.health.qa.autotest.plr.fhir.model.PlrFhirResourceType;
-import ca.bc.gov.health.qa.autotest.plr.fhir.maintain.FacilityQueryResponseMapper;
-import ca.bc.gov.health.qa.autotest.plr.fhir.maintain.MaintainFacilityBuilder;
-import ca.bc.gov.health.qa.autotest.plr.fhir.maintain.MaintainOrgBuilder;
-import ca.bc.gov.health.qa.autotest.plr.fhir.maintain.OrgQueryResponseMapper;
 import ca.bc.gov.health.qa.autotest.plr.util.UserType;
 import ca.bc.gov.health.qa.autotest.runner.util.log.ExecutionLogManager;
