+++ autotest-plr/src/main/java/ca/bc/gov/health/qa/autotest/plr/fhir/data/facility/FacilityBuilderFactory.java
@@ -1,6 +1,6 @@
-package ca.bc.gov.health.qa.autotest.plr.fhir.data;
+package ca.bc.gov.health.qa.autotest.plr.fhir.data.facility;
 
-import ca.bc.gov.health.qa.autotest.plr.fhir.maintain.MaintainFacilityBuilder;
+import ca.bc.gov.health.qa.autotest.plr.fhir.maintain.facility.MaintainFacilityBuilder;
 
 /**
  * Factory responsible for creating and configuring {@link MaintainFacilityBuilder}
