+++ autotest-plr/src/main/java/ca/bc/gov/health/qa/autotest/plr/fhir/maintain/facility/FacilityQueryResponseMapper.java
@@ -1,7 +1,5 @@
-package ca.bc.gov.health.qa.autotest.plr.fhir.maintain;
+package ca.bc.gov.health.qa.autotest.plr.fhir.maintain.facility;
 
-import ca.bc.gov.health.qa.autotest.runner.util.log.ExecutionLogManager;
-import org.apache.logging.log4j.Logger;
 import org.json.JSONArray;
 import org.json.JSONObject;
 
@@ -14,8 +12,6 @@ import ca.bc.gov.health.qa.autotest.plr.fhir.model.IdentifierType;
  * caught by MaintainFacilityBuilder.verifyParameters() on build().
  */
 public final class FacilityQueryResponseMapper {
-
-    private static final Logger LOG = ExecutionLogManager.getLogger();
 
     private FacilityQueryResponseMapper() {}
