+++ autotest-plr/src/main/java/ca/bc/gov/health/qa/autotest/plr/fhir/maintain/facility/MaintainFacilityBuilder.java
@@ -1,10 +1,13 @@
-package ca.bc.gov.health.qa.autotest.plr.fhir.maintain;
+package ca.bc.gov.health.qa.autotest.plr.fhir.maintain.facility;
 
+import ca.bc.gov.health.qa.autotest.plr.fhir.data.facility.FacilityAttribute;
+import ca.bc.gov.health.qa.autotest.plr.fhir.maintain.MaintainRequestBuilder;
+import ca.bc.gov.health.qa.autotest.plr.fhir.maintain.common.MaintainAccessor;
+import ca.bc.gov.health.qa.autotest.plr.fhir.maintain.common.MaintainUtils;
 import ca.bc.gov.health.qa.autotest.plr.fhir.model.EndReasonCode;
 
 import static java.util.Objects.requireNonNull;
 
-import java.lang.invoke.MethodHandles;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -13,8 +16,6 @@ import java.util.Map;
 import org.json.JSONArray;
 import org.json.JSONObject;
 
-import ca.bc.gov.health.qa.autotest.core.util.io.ResourceUtils;
-import ca.bc.gov.health.qa.autotest.plr.fhir.data.FacilityAttribute;
 import ca.bc.gov.health.qa.autotest.plr.fhir.model.IdentifierType;
 import ca.bc.gov.health.qa.autotest.plr.fhir.model.PlrFhirResourceType;
 
@@ -149,9 +150,7 @@ public class MaintainFacilityBuilder implements MaintainRequestBuilder
     public JSONObject build()
     {
         verifyParameters();
-        String template = ResourceUtils.readResource(
-                MethodHandles.lookup().lookupClass(), "maintain-facility.json");
-        JSONObject json = new JSONObject(template);
+        JSONObject json = MaintainUtils.readJsonTemplate("maintain-facility.json");
 
         MaintainAccessor accessor = new MaintainAccessor(json);
         JSONObject facilityJson = accessor.getFacilityJson();
