+++ autotest-plr/src/main/java/ca/bc/gov/health/qa/autotest/plr/fhir/maintain/organization/MaintainOrgBuilder.java
@@ -1,8 +1,7 @@
-package ca.bc.gov.health.qa.autotest.plr.fhir.maintain;
+package ca.bc.gov.health.qa.autotest.plr.fhir.maintain.organization;
 
 import static java.util.Objects.requireNonNull;
 
-import java.lang.invoke.MethodHandles;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -11,12 +10,14 @@ import java.util.Map;
 import org.json.JSONArray;
 import org.json.JSONObject;
 
-import ca.bc.gov.health.qa.autotest.core.util.io.ResourceUtils;
 import ca.bc.gov.health.qa.autotest.plr.fhir.model.OrgRoleType;
 import ca.bc.gov.health.qa.autotest.plr.fhir.model.IdentifierType;
 import ca.bc.gov.health.qa.autotest.plr.fhir.model.PlrFhirResourceType;
+import ca.bc.gov.health.qa.autotest.plr.fhir.data.organization.OrganizationAttribute;
+import ca.bc.gov.health.qa.autotest.plr.fhir.maintain.MaintainRequestBuilder;
+import ca.bc.gov.health.qa.autotest.plr.fhir.maintain.common.MaintainAccessor;
+import ca.bc.gov.health.qa.autotest.plr.fhir.maintain.common.MaintainUtils;
 import ca.bc.gov.health.qa.autotest.plr.fhir.model.HdsType;
-import ca.bc.gov.health.qa.autotest.plr.fhir.data.OrganizationAttribute;
 
 /**
  * Builder for Organization maintain requests. Supports configuration of multi-valued
@@ -139,9 +140,7 @@ public class MaintainOrgBuilder implements MaintainRequestBuilder
     public JSONObject build()
     {
         verifyParameters();
-        String template = ResourceUtils.readResource(
-                MethodHandles.lookup().lookupClass(), "maintain-organization.json");
-        JSONObject json = new JSONObject(template);
+        JSONObject json = MaintainUtils.readJsonTemplate("maintain-organization.json");
 
         MaintainAccessor accessor = new MaintainAccessor(json);
         JSONObject orgJson = accessor.getOrgJson();
