+++ autotest-plr/src/main/java/ca/bc/gov/health/qa/autotest/plr/web/pages/plr/facility/UpdateFacilityPage.java
@@ -5,7 +5,7 @@ import static org.testng.Assert.fail;
 import java.net.URI;
 import java.util.Map;
 
-import ca.bc.gov.health.qa.autotest.plr.fhir.maintain.MaintainOrgBuilder;
+import ca.bc.gov.health.qa.autotest.plr.fhir.maintain.organization.MaintainOrgBuilder;
 import ca.bc.gov.health.qa.autotest.plr.web.pages.components.AutocompleteMenu;
 import ca.bc.gov.health.qa.autotest.plr.web.pages.components.DropDownMenu;
 import org.apache.commons.lang3.StringUtils;
